Task_Movie_Rental/requirements.txt
fastapi==0.115.6
jose==1.0.0
passlib==1.7.4
pydantic==2.10.4
python-dotenv==1.0.1
python_jose==3.3.0
pytz==2024.2
starlette==0.42.0
uvicorn==0.34.0
